change cp1 to use double precision

+++ cp1/cp.cc
@@ -1,8 +1,8 @@
 #include <cmath>
 
 float correlation(int row_i, int row_j, const float *data, int row_width);
-float get_data(int x, int y, const float *data, int nx);
-float get_deno_term(int n, float sq_sum, float sum);
+double get_data(int x, int y, const float *data, int nx);
+double get_deno_term(int n, double sq_sum, double sum);
 
 /*
 This is the function you need to implement. Quick reference:
@@ -29,16 +29,16 @@ void correlate(int ny, int nx, const float *data, float *result) {
  * code relates to the row/col coordinates. Our samples are the rows.
  * */
 float correlation(int row_i, int row_j, const float *data, int row_width) {
-  float row_dot = 0;
-  float row_i_sum = 0;
-  float row_j_sum = 0;
+  double row_dot = 0;
+  double row_i_sum = 0;
+  double row_j_sum = 0;
 
-  float row_i_sq_sum = 0;
-  float row_j_sq_sum = 0;
+  double row_i_sq_sum = 0;
+  double row_j_sq_sum = 0;
 
   for (int col=0; col < row_width; col++) {
-    float row_i_col = get_data(col, row_i, data, row_width);
-    float row_j_col = get_data(col, row_j, data, row_width);
+    double row_i_col = get_data(col, row_i, data, row_width);
+    double row_j_col = get_data(col, row_j, data, row_width);
 
     row_dot += row_i_col * row_j_col;
 
@@ -49,18 +49,18 @@ float correlation(int row_i, int row_j, const float *data, int row_width) {
     row_j_sq_sum += row_j_col * row_j_col;
   }
 
-  float numerator = (row_width * row_dot) - (row_i_sum * row_j_sum);
-  float deno_row_i_term = get_deno_term(row_width, row_i_sq_sum, row_i_sum);
-  float deno_row_j_term = get_deno_term(row_width, row_j_sq_sum, row_j_sum);
+  double numerator = (row_width * row_dot) - (row_i_sum * row_j_sum);
+  double deno_row_i_term = get_deno_term(row_width, row_i_sq_sum, row_i_sum);
+  double deno_row_j_term = get_deno_term(row_width, row_j_sq_sum, row_j_sum);
 
   return numerator / (deno_row_i_term * deno_row_j_term);
 }
 
-float get_data(int x, int y, const float *data, int nx) {
+double get_data(int x, int y, const float *data, int nx) {
   return data[x + y*nx];
 }
 
-float get_deno_term(int n, float sq_sum, float sum) {
-  float sum_squared = sum * sum;
+double get_deno_term(int n, double sq_sum, double sum) {
+  double sum_squared = sum * sum;
   return std::sqrt((n * sq_sum) - sum_squared);
 }
